requirements.txt
numpy==1.26.3
Pillow==10.1.0
pygame==2.5.2
